WI/requirements.txt
flask
discord.py
python-dotenv
better-profanity
gunicorn